+++ _examples/view/parse-template/main.go
@@ -1,22 +1,28 @@
 // Package main shows how to parse a template through custom byte slice content.
+// The following works with HTML, Pug and Ace template parsers.
+// To learn how you can manually parse a template from a text for the rest
+// template parsers navigate through the example's subdirectories.
 package main
 
 import "github.com/kataras/iris/v12"
 
 func main() {
-	app := iris.New()
 	// To not load any templates from files or embedded data,
 	// pass nil or empty string on the first argument:
-	// view := iris.HTML(nil, ".html")
+	// e := iris.HTML(nil, ".html")
 
-	view := iris.HTML("./views", ".html")
-	view.ParseTemplate("program.html", []byte(`<h1>{{greet .Name}}</h1>`), iris.Map{
+	e := iris.HTML("./views", ".html")
+	// e := iris.Pug("./views",".pug")
+	// e := iris.Ace("./views",".ace")
+	e.ParseTemplate("program.html", []byte(`<h1>{{greet .Name}}</h1>`), iris.Map{
 		"greet": func(name string) string {
 			return "Hello, " + name + "!"
 		},
 	})
 
-	app.RegisterView(view)
+	app := iris.New()
+	app.RegisterView(e)
+
 	app.Get("/", index)
 	app.Get("/layout", layout)